tests/wpt/web-platform-tests/notifications/event-onshow.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Notification.onshow (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
    if (Notification.permission != "granted") {
        this.force_timeout()
        this.set_status(this.NOTRUN, "You must allow notifications for this"
            + " origin before running this test.")
    } else {
        var t = async_test("Invoked the onshow event handler.")
        var notification = new Notification("New Email Received")
        notification.onshow = t.step_func(function(e) {
            assert_equals(Object.prototype.toString.call(e), "[object Event]")
            assert_equals(e.type, "show", "Checked the event type.")
            notification.close()
            t.done()
        })
    }
}, "Checked test prerequisites.")
</script>
